+++ problems/palindrome-number/languages/solution.ts
@@ -1,0 +1,7 @@
+function isPalindrome(x: number): boolean {
+    // convert number to string;
+    const str = x.toString();
+
+    // compare with reversed version
+    return str === str.split('').reverse().join('');
+};